LPS-72199 portal-search-solr: Split into Pagination and Count
https://issues.liferay.com/browse/LPS-72199
